+++ src/main/java/me/desair/tus/server/download/DownloadGetRequestHandler.java
@@ -9,13 +9,12 @@ import me.desair.tus.server.HttpHeader;
 import me.desair.tus.server.HttpMethod;
 import me.desair.tus.server.ProtocolVersion;
 import me.desair.tus.server.exception.TusException;
-import me.desair.tus.server.exception.UploadInProgressException;
+import me.desair.tus.server.exception.UploadNotFoundException;
 import me.desair.tus.server.upload.UploadInfo;
 import me.desair.tus.server.upload.UploadStorageService;
 import me.desair.tus.server.util.AbstractRequestHandler;
 import me.desair.tus.server.util.TusServletRequest;
 import me.desair.tus.server.util.TusServletResponse;
-import me.desair.tus.server.util.Utils;
 
 /** Send the uploaded bytes of finished uploads. */
 public class DownloadGetRequestHandler extends AbstractRequestHandler {
@@ -30,8 +29,7 @@ public class DownloadGetRequestHandler extends AbstractRequestHandler {
 
   @Override
   public boolean supports(HttpMethod method, ProtocolVersion version) {
-    return HttpMethod.GET.equals(method)
-        && (version == ProtocolVersion.TUS_1_0_0 || version == ProtocolVersion.RUFH);
+    return supports(method);
   }
 
   @Override
@@ -43,35 +41,32 @@ public class DownloadGetRequestHandler extends AbstractRequestHandler {
       String ownerKey)
       throws IOException, TusException {
 
-    ProtocolVersion version = Utils.detectProtocolVersion(servletRequest, ProtocolVersion.AUTO);
     UploadInfo info = uploadStorageService.getUploadInfo(servletRequest.getRequestURI(), ownerKey);
-    if (version == ProtocolVersion.RUFH && info != null && info.isUploadInProgress()) {
+    if (info == null || info.isExpired()) {
+
+      throw new UploadNotFoundException("The requested upload cannot be found or has expired");
+    }
+
+    if (info.isUploadInProgress()) {
       // Delegate to RufhHeadGetRequestHandler for RUFH offset retrieval on in-progress uploads
+      servletResponse.setStatus(HttpServletResponse.SC_NO_CONTENT);
+      servletResponse.setHeader(HttpHeader.CONTENT_LENGTH, "0");
       return;
     }
 
-    if (info == null || info.isUploadInProgress() || info.isExpired()) {
-      throw new UploadInProgressException(
-          "Upload "
-              + servletRequest.getRequestURI()
-              + " is still in progress "
-              + "and cannot be downloaded yet");
-    } else {
+    servletResponse.setHeader(HttpHeader.CONTENT_LENGTH, Objects.toString(info.getLength()));
 
-      servletResponse.setHeader(HttpHeader.CONTENT_LENGTH, Objects.toString(info.getLength()));
+    servletResponse.setHeader(
+        HttpHeader.CONTENT_DISPOSITION,
+        String.format(
+            CONTENT_DISPOSITION_FORMAT,
+            info.getFileName().replace("\"", ""),
+            URLEncoder.encode(info.getFileName(), StandardCharsets.UTF_8.toString())
+                .replace("+", "%20")));
 
-      servletResponse.setHeader(
-          HttpHeader.CONTENT_DISPOSITION,
-          String.format(
-              CONTENT_DISPOSITION_FORMAT,
-              info.getFileName().replace("\"", ""),
-              URLEncoder.encode(info.getFileName(), StandardCharsets.UTF_8.toString())
-                  .replace("+", "%20")));
+    servletResponse.setHeader(HttpHeader.CONTENT_TYPE, info.getFileMimeType());
 
-      servletResponse.setHeader(HttpHeader.CONTENT_TYPE, info.getFileMimeType());
-
-      uploadStorageService.copyUploadTo(info, servletResponse.getOutputStream());
-    }
+    uploadStorageService.copyUploadTo(info, servletResponse.getOutputStream());
 
     servletResponse.setStatus(HttpServletResponse.SC_OK);
   }