+++ src/main/java/me/desair/tus/server/rufh/handler/RufhUploadLimitHeaderRequestHandler.java
@@ -47,11 +47,17 @@ public class RufhUploadLimitHeaderRequestHandler extends AbstractRequestHandler 
     }
 
     String uploadUri = servletResponse.getHeader(HttpHeader.LOCATION);
-    if (StringUtils.isBlank(uploadUri)) {
+    if (StringUtils.isBlank(uploadUri) && servletRequest != null) {
       uploadUri = servletRequest.getRequestURI();
     }
 
-    UploadInfo uploadInfo = uploadStorageService.getUploadInfo(uploadUri, ownerKey);
+    UploadInfo uploadInfo = null;
+    try {
+      uploadInfo = uploadStorageService.getUploadInfo(uploadUri, ownerKey);
+    } catch (Exception e) {
+      uploadInfo = null;
+    }
+
     addUploadLimitHeader(servletResponse, uploadStorageService, uploadInfo);
   }